+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 0.7.0 
 - [Enhancement] Added ability to verify signed Stripe Webhook payloads via `Stripe.Webhook.construct_event/4` (Thanks @jayjun)
-- [Enhancement] Added ability to generate a Connect callback url via `Stripe.Connect.generate_button_url`
+- [Enhancement] Added ability to generate a Connect callback url via `Stripe.Connect.generate_button_url/2`